requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
Jinja2==3.1.3
numpy==1.24.3
python-dotenv==1.0.1
Flask-WTF==1.2.1
WTForms==3.1.2 